Move ERU owner test
